Added content to the about me page and testing map
Added responsive content in the about me section and created the resources cards. I used flexbox for the cards but need to look into how to get them to wrap evenly. Still need to work on adding links and hover effects. Began playing with the amcharts map plugin but need to look into more methods to style the map.