Began adding elements to the map
Got elements to show up in the description when map objects are clicked. Need to work on styling the descriptions and changing the hover effects. Tried working on the auto zoom to follow the animation line path but need to work on the timing functions. May need to put the auto zoom in the icebox until rest of site is complete.