Added more gallery pages and maps
Reorganized the file layout and began working on gallery pages. Got the main gallery page layout close to finished and started working on the trips pages. I have the multiple maps working with a crude navigation to switch between the two. Still need to decide how to display the photos when a portion of the map is selected. May want to put an image slider and autoscroll to help users find the content. Also need to do more styling on the page and figure out how I will link to the other galleries.